requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
asyncer==0.0.7
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
banks==2.1.2
beautifulsoup4==4.13.4
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chainlit==2.5.5
charset-normalizer==3.4.2
chevron==0.14.0
click==8.2.1
click-default-group==1.2.4
colorama==0.4.6
comm==0.2.2
condense-json==0.1.3
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
greenlet==3.2.3
griffe==1.7.3
grpcio==1.73.0
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
ipykernel==6.29.5
ipython==8.37.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
Lazify==0.4.0
literalai==0.1.201
llama-cloud==0.1.26
llama-cloud-services==0.6.32
llama-index==0.12.42
llama-index-agent-openai==0.4.11
llama-index-cli==0.4.3
llama-index-core==0.12.42
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.7
llama-index-llms-openai==0.4.6
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.9
llama-index-readers-llama-parse==0.4.0
llama-index-readers-wikipedia==0.3.0
llama-parse==0.6.32
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.9.4
mistune==3.1.3
monotonic==1.6
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.42.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.4.3
notebook_shim==0.2.4
numpy==2.2.6
openai==1.86.0
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-alephalpha==0.40.9
opentelemetry-instrumentation-anthropic==0.40.9
opentelemetry-instrumentation-bedrock==0.40.9
opentelemetry-instrumentation-chromadb==0.40.9
opentelemetry-instrumentation-cohere==0.40.9
opentelemetry-instrumentation-crewai==0.40.9
opentelemetry-instrumentation-google-generativeai==0.40.9
opentelemetry-instrumentation-groq==0.40.9
opentelemetry-instrumentation-haystack==0.40.9
opentelemetry-instrumentation-lancedb==0.40.9
opentelemetry-instrumentation-langchain==0.40.9
opentelemetry-instrumentation-llamaindex==0.40.9
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-instrumentation-marqo==0.40.9
opentelemetry-instrumentation-mcp==0.40.9
opentelemetry-instrumentation-milvus==0.40.9
opentelemetry-instrumentation-mistralai==0.40.9
opentelemetry-instrumentation-ollama==0.40.9
opentelemetry-instrumentation-openai==0.40.9
opentelemetry-instrumentation-pinecone==0.40.9
opentelemetry-instrumentation-qdrant==0.40.9
opentelemetry-instrumentation-replicate==0.40.9
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-sagemaker==0.40.9
opentelemetry-instrumentation-sqlalchemy==0.52b1
opentelemetry-instrumentation-threading==0.52b1
opentelemetry-instrumentation-together==0.40.9
opentelemetry-instrumentation-transformers==0.40.9
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-vertexai==0.40.9
opentelemetry-instrumentation-watsonx==0.40.9
opentelemetry-instrumentation-weaviate==0.40.9
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-semantic-conventions-ai==0.4.9
opentelemetry-util-http==0.52b1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
posthog==3.25.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.29
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pypdf==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.2
python-json-logger==3.3.0
python-multipart==0.0.18
python-socketio==5.13.0
python-ulid==3.0.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
scikit-learn==1.7.0
scipy==1.15.3
Send2Trash==1.8.3
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlite-fts4==1.0.3
sqlite-migrate==0.1b0
sqlite-utils==3.38
sse-starlette==2.3.6
stack-data==0.6.3
starlette==0.41.3
streamlit==1.45.1
striprtf==0.0.26
syncer==2.0.3
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
traceloop-sdk==0.40.9
traitlets==5.14.3
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uptrace==1.31.0
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.3
watchfiles==0.20.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
